episode_generator_requirements.txt
torch
python-dotenv
flask
pydub
colorama
jinja2
requests
TTS
pillow